+++ src/Logger/Maintenance.cs
@@ -1,6 +1,6 @@
 ﻿/* PROJECT: Logger (https://github.com/aprettycoolprogram/Logger)
  *    FILE: Logger.Maintenance.cs
- * UPDATED: 7-1-2021-8:45 PM
+ * UPDATED: 7-7-2021-11:47 AM
  * LICENSE: Apache v2 (https://apache.org/licenses/LICENSE-2.0)
  *          Copyright 2021 A Pretty Cool Program All rights reserved
  */
@@ -26,8 +26,11 @@ namespace Logger
             {
                 Directory.CreateDirectory(directoryPath);
 
-                Timestamped.LogEvent("system", "SYSTEM", Assembly.GetExecutingAssembly().GetName().Name, $"Created directory: {directoryPath}");
+                LogEvent.Timestamped("system", "SYSTEM", Assembly.GetExecutingAssembly().GetName().Name, "switch(mawsCommand) case: InptAdmitDate [{mawsCommand}]");
             }
         }
     }
 }
+
+/* Why hardcoded
+ */