+++ src/MyAvatoolWebService.asmx.cs
@@ -1,19 +1,18 @@
 ﻿/* PROJECT: MyAvatoolWebService (https://github.com/aprettycoolprogram/MyAvatoolWebService)
  *    FILE: MyAvatoolWebService.MyAvatoolWebService.asmx.cs
- * UPDATED: 7-6-2021-4:12 PM
+ * UPDATED: 7-7-2021-11:55 AM
  * LICENSE: Apache v2 (https://apache.org/licenses/LICENSE-2.0)
  *          Copyright 2021 A Pretty Cool Program All rights reserved
  */
 
-/* MyAvatoolWebService main class.
- *
- * Development notes/comments can be found at the end of this class.
+/* Entry point for MAWS.
  */
 
 using System.Collections.Generic;
 using System.Reflection;
 using System.Web.Services;
 using NTST.ScriptLinkService.Objects;
+using Logger;
 
 namespace MyAvatoolWebService
 {
@@ -25,7 +24,7 @@ namespace MyAvatoolWebService
     [System.ComponentModel.ToolboxItem(false)]
     public class MyAvatoolWebService : WebService
     {
-        public Dictionary<string, string> MawsSetting;
+        //public Dictionary<string, string> MawsSetting;
 
         /// <summary>
         /// Returns the MAWS version.
@@ -34,12 +33,8 @@ namespace MyAvatoolWebService
         [WebMethod]
         public string GetVersion()
         {
-            //MawsSetting = Settings.GetSettings();
-
-            //if(MawsSetting["TestFunctionality"] == "true")
-            //{
-            //    ForceTest(MawsSetting);
-            //}
+            // This line has to be commented out in production, otherwise MAWS will not work!
+            ForceTest();
 
             return "VERSION 1.0";
         }
@@ -53,71 +48,66 @@ namespace MyAvatoolWebService
         [WebMethod]
         public OptionObject2015 RunScript(OptionObject2015 sentOptionObject, string mawsRequest)
         {
-            MawsSetting = Settings.GetSettings();
-
-            Logger.Timestamped.LogEvent(MawsSetting["Logging"].ToLower(), "TRACE", Assembly.GetExecutingAssembly().GetName().Name, $"Initial MAWS Request: {mawsRequest}");
+            Dictionary<string, string> MawsSetting = Settings.GetSettings();
+            var logSetting                         = MawsSetting["Logging"].ToLower();
+            var assemblyName                       = Assembly.GetExecutingAssembly().GetName().Name;
+            LogEvent.Timestamped(logSetting, "TRACE", assemblyName, $"Initial MAWS Request: {mawsRequest}");
 
             var mawsCommand = RequestSyntaxEngine.RequestComponent.GetCommand(mawsRequest);
-            Logger.Timestamped.LogEvent(MawsSetting["Logging"].ToLower(), "TRACE", Assembly.GetExecutingAssembly().GetName().Name, $"Initial MAWS Command to be executed: {mawsCommand}");
+            LogEvent.Timestamped(logSetting, "TRACE", assemblyName, $"Initial MAWS Command: {mawsCommand}");
 
             OptionObject2015 completedOptionObject;
 
             switch(mawsCommand)
             {
                 case "inptadmitdate":
-                    Logger.Timestamped.LogEvent(MawsSetting["Logging"].ToLower(), "TRACE", Assembly.GetExecutingAssembly().GetName().Name, "case: inptadmitdate");
-                    completedOptionObject = InptAdmitDate.Execute.Action(sentOptionObject, mawsRequest);
+                    LogEvent.Timestamped(logSetting, "TRACE", assemblyName, "switch(mawsCommand) case: InptAdmitDate [{mawsCommand}]");
+                    //completedOptionObject = InptAdmitDate.Execute.Action(sentOptionObject, mawsRequest);
                     break;
 
                 case "dose":
-                    Logger.Timestamped.LogEvent(MawsSetting["Logging"].ToLower(), "TRACE", Assembly.GetExecutingAssembly().GetName().Name, "case: dose");
-                    completedOptionObject = Dose.Execute.Action(sentOptionObject, mawsRequest);
+                    LogEvent.Timestamped(logSetting, "TRACE", assemblyName, "switch(mawsCommand) case: Dose [{mawsCommand}]");
+                    //completedOptionObject = Dose.Execute.Action(sentOptionObject, mawsRequest);
                     break;
 
                 case "newdevelopment":
-                    Logger.Timestamped.LogEvent(MawsSetting["Logging"].ToLower(), "TRACE", Assembly.GetExecutingAssembly().GetName().Name, "case: newdevelopemnt");
-                    completedOptionObject = NewDevelopment.Execute.Action(sentOptionObject, mawsRequest);
+                    LogEvent.Timestamped(logSetting, "TRACE", assemblyName, "switch(mawsCommand) case: NewDevelopment [{mawsCommand}]");
+                    //completedOptionObject = NewDevelopment.Execute.Action(sentOptionObject, mawsRequest);
                     break;
 
                 default:
-                    Logger.Timestamped.LogEvent(MawsSetting["Logging"].ToLower(), "ERROR", Assembly.GetExecutingAssembly().GetName().Name, $"Invalid MAWS Command: \"{mawsCommand}\".");
-                    completedOptionObject = sentOptionObject;
+                    LogEvent.Timestamped(logSetting, "ERROR", assemblyName, $"Invalid MAWS Command: \"{mawsCommand}\".");
+                    //completedOptionObject = sentOptionObject;
                     break;
             }
+
+            completedOptionObject = new OptionObject2015(); // TEMP TESTING
 
             return completedOptionObject;
         }
 
         /// <summary>
-        /// Force a bunch of tests (if testing is enabled in MAWS.settings
+        /// Force a bunch of MAWS functionality tests.
         /// </summary>
-        /// <param name="mawsSettings">The MAWS settings.</param>
-        public void ForceTest(Dictionary<string, string> mawsSettings)
+        public void ForceTest()
         {
-            if(mawsSettings["TestMaws"] == "true")
+            // Trace commands/actions/options.
+            var emptyOptionObject = new OptionObject2015();
+            _ = RunScript(emptyOptionObject, "InptAdmitDate-action-option");
+            _ = RunScript(emptyOptionObject, "Dose-action-option");
+            _ = RunScript(emptyOptionObject, "NewDevelopment-action-option");
+            _ = RunScript(emptyOptionObject, "Fake-action-option");
+
+            // Test against an partially initialized OptionObject.
+            var testInptAdmitDateOptionObject= new OptionObject2015
             {
-                {
-                    // Test against an empty OptionObject.
-                    var emptyOptionObject = new OptionObject2015();
-                    _ = RunScript(emptyOptionObject, "InptAdmitDate-action-option");
-                    _ = RunScript(emptyOptionObject, "Dose-action-option");
-                    _ = RunScript(emptyOptionObject, "NewDevelopment-action-option");
-                    _ = RunScript(emptyOptionObject, "Fake-action-option");
+                ErrorCode = 0,
+                ErrorMesg = "",
+            };
+            _ = RunScript(testInptAdmitDateOptionObject, "InptAdmitDate-ComparePreAdmitToAdmit");
 
-                    // Test against an initialized OptionObject.
-                    var testInptAdmitDateOptionObject= new OptionObject2015
-                    {
-                        ErrorCode = 0,
-                        ErrorMesg = "",
-                    };
-                    _ = RunScript(testInptAdmitDateOptionObject, "InptAdmitDate-ComparePreAdmitToAdmit");
-                }
-
-                if(mawsSettings["TestRequestSyntaxEngine"] == "true")
-                {
-                    RequestSyntaxEngine.TestFunctionality.Force();
-                }
-            }
+            // Test RequestSyntaxEngine functionality.
+            RequestSyntaxEngine.TestFunctionality.Force();
         }
     }
 }
@@ -130,16 +120,18 @@ namespace MyAvatoolWebService
  *
  * - Both GetVersion() and RunScript() are required by myAvatar, so don't remove them.
  *
+ * - The "VERSION" will always be the target version that is being developed. For example, "VERSION 1.0" during v1.0
+ *   development, "VERSION 1.1" during v1.1 development.
+ *
  * - ForceTest() is here because otherwise it's a pain to get working, but eventually I would like to move it into its
  *   own class.
- *
- * - I'm leaving the "VERSION" as "1.0" throughout development of v1.0.
  *
  * ------------
  * GetVersion()
  * ------------
- * - Injecting code into GetVersion() is probably a Bad Idea, but I wanted an easy way to test functionality without
- *   having to publish the web service every time I made a change.
+ * - Injecting code into GetVersion() is a Bad Idea, and if you don't comment-out the `ForceTest()` line, MAWS won't
+ *   work. However I want an easy way to test some functionality on my local machine without having to publish the web
+ *   service. This functionality will probably be depreciated further down development.
  *
  *   Here is how you can test MAWS:
  *      1. Make sure you have the proper testing settings configured in maws.settings